feat: add state and dispatch mapping functions to book component file
