+++ src/components/book.js
@@ -1,5 +1,8 @@
 import React from 'react';
+import { connect } from 'react-redux';
 import Proptypes from 'prop-types';
+import { createAction, removeAction } from '../actions/index';
+
 
 const Book = props => {
   const { id, title, category } = props;
@@ -11,10 +14,19 @@ const Book = props => {
     </tr>
   );
 };
-export default Book;
 
 Book.propTypes = {
   id: Proptypes.number.isRequired,
   title: Proptypes.string.isRequired,
   category: Proptypes.string.isRequired,
 };
+
+const mapStateToProps = state => ({
+  books: state,
+});
+const mapDispatchToProps = dispatch => ({
+  addBook: () => dispatch(createAction()),
+  removeBook: () => dispatch(removeAction()),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(Book);
