+++ src/reducers/books.js
@@ -1,10 +1,32 @@
 import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
+import randomNumber from '../logic';
 
-const booksReducer = (state, action) => {
+const initialState = {
+  books: [
+    {
+      id: randomNumber(),
+      title: 'css',
+      category: 'Programming',
+    },
+    {
+      id: randomNumber(),
+      title: 'Punctuation',
+      category: 'English',
+    },
+    {
+      id: randomNumber(),
+      title: 'Polynomials',
+      category: 'Mathematics',
+    },
+  ],
+};
+
+const booksReducer = (state = initialState, action) => {
   switch (action.type) {
     case CREATE_BOOK:
       return [
         ...state,
+        action.book,
       ];
     case REMOVE_BOOK:
       return [
